requirements.txt
pyodbc==5.2.0
pandas==2.2.0
SQLAlchemy==2.0.42
python-dotenv==1.0.1 